refactor: 콘서트 및 페스티벌 API의 폼 데이터 생성 로직을 utils로 분리
